rmlui: publish version 6.2, unvendor deps, stop publishing revs for older versions, modernized (#30498)
* rmlui: add initial support for 5.x and partial for 6.x

* rmlui: fix patches and support for 6.x

* rmlui: rename 4.x to all folder

* Leave only latest version and correctly unvendor dependencies

* Simplify test_package

* Fix windows shared build

* Remove vendored headers and drop with_thirdparty_containers option

* Only package needed licenses

* Fix case sensitive removal on Linux

---------
