--- recipes/rmlui/3.3/test_package/conanfile.py
@@ -1,18 +1,0 @@
-import os
-
-from conans import ConanFile, CMake, tools
-
-
-class ConanRmluiTestConan(ConanFile):
-    settings = "os", "compiler", "build_type", "arch"
-    generators = "cmake"
-
-    def build(self):
-        cmake = CMake(self)
-        cmake.configure()
-        cmake.build()
-
-    def test(self):
-        if not tools.cross_building(self):
-            bin_path = os.path.join("bin", "example")
-            self.run(bin_path, run_environment=True)
