--- recipes/rmlui/3.3/test_package/example.cpp
@@ -1,7 +1,0 @@
-#include <RmlUi/Core.h>
-
-int main()
-{
-    Rml::Core::Shutdown();
-    return 0;
-}
