--- recipes/rmlui/4.x/test_v1_package/conanfile.py
@@ -1,18 +1,0 @@
-from conans import ConanFile, CMake, RunEnvironment, tools
-from conan.tools.build import cross_building
-import os
-
-# legacy validation with Conan 1.x
-class ConanRmluiTestV1Conan(ConanFile):
-    settings = "os", "arch", "compiler", "build_type"
-    generators = "cmake", "cmake_find_package_multi"
-
-    def build(self):
-        cmake = CMake(self)
-        cmake.configure()
-        cmake.build()
-
-    def test(self):
-        if not cross_building(self):
-            bin_path = os.path.join("bin", "test_package")
-            self.run(bin_path, run_environment=True)
