+++ recipes/rmlui/all/conandata.yml
@@ -1,0 +1,7 @@
+sources:
+  "6.2":
+    url: "https://github.com/mikke89/RmlUi/archive/refs/tags/6.2.tar.gz"
+    sha256: "814c3ff7b9666280338d8f0dda85979f5daf028d01c85fc8975431d1e2fd8e8b"
+patches:
+  "6.2":
+    - patch_file: "patches/6.2-unvendor-deps.patch"
