+++ recipes/rmlui/all/test_package/conanfile.py
@@ -7,13 +7,9 @@ import os
 class ConanRmluiTestConan(ConanFile):
     settings = "os", "arch", "compiler", "build_type"
     generators = "CMakeToolchain", "CMakeDeps", "VirtualRunEnv"
-    test_type = "explicit"
 
     def requirements(self):
         self.requires(self.tested_reference_str)
-
-    def configure(self):
-        self.options[self.tested_reference_str].with_thirdparty_containers = True
 
     def layout(self):
         cmake_layout(self)
@@ -25,5 +21,5 @@ class ConanRmluiTestConan(ConanFile):
 
     def test(self):
         if can_run(self):
-            bin_path = os.path.join(self.cpp.build.bindirs[0], "test_package")
+            bin_path = os.path.join(self.cpp.build.bindir, "test_package")
             self.run(bin_path, env="conanrun")
