+++ recipes/rmlui/all/test_package/example.cpp
@@ -1,0 +1,6 @@
+#include <RmlUi/Core.h>
+
+int main() {
+  Rml::Initialise();
+  Rml::Shutdown();
+}
