+++ recipes/rmlui/config.yml
@@ -1,13 +1,3 @@
 versions:
-  "3.3":
-    folder: "3.3"
-  "4.0":
-    folder: "4.x"
-  "4.1":
-    folder: "4.x"
-  "4.2":
-    folder: "4.x"
-  "4.3":
-    folder: "4.x"
-  "4.4":
-    folder: "4.x"
+  "6.2":
+    folder: "all"
